# api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: CDE Platform engineering
  description: |
    Common Data Environment API for construction and asset information.

    ## Tenant scoping

    Every request is scoped to exactly one tenant, and the tenant is taken **from the authenticated principal** — never from a path, query, header, or body parameter. There is no tenant selector on any endpoint, and supplying one has no effect. A resource belonging to another tenant is not visible: reading it by id returns `404`, not `403`, so the API does not confirm that an id exists somewhere else.

    ## Errors

    Every error response is an RFC 9457 Problem Details document with the media type `application/problem+json`. It carries a `traceId` that identifies the request in the server logs — quote it when reporting a problem. The `type` member is a stable relative URI identifying the kind of failure, and is safe to branch on when the status code alone is too coarse; `detail` is human-readable text whose wording may change, and is not.

    Validation failures return `422` with an `invalidFields` array naming each rejected field. A request body that cannot be parsed at all returns `400`, because no field can be named.

    ## Rate limits

...

    ## Tracing

    Every response carries `X-Trace-Id`. Send a W3C `traceparent` header and its trace id is adopted, so a client-side trace and the server-side logs share one identifier.
  license:
    name: Proprietary
  title: CDE Platform API
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /api/viewer/converter-status:
    get:
      description: |-
        The viewer asks before offering conversion-dependent actions, so it can say why something is unavailable rather than failing when someone tries it.

        Requires authentication.
      operationId: getConverterStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConverterStatusResponse"
                example: null
          description: Whether the conversion service answered.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Check whether the conversion service is reachable
      tags:
      - Viewer
  /api/viewer/{documentId}:
    get:
      description: |-
        Returns whatever suits the document's format: SVG markup for a drawing, a pointer to the bytes for a PDF, extracted geometry for a model, or an explanation of why it cannot be opened. Read the `type` member first — it decides the shape of the rest.

        Images and mesh formats (GLB, STL, OBJ, PLY, DAE) are returned as raw bytes with the matching content type rather than as JSON, so check the response's content type before parsing it. Office documents are converted and returned as PDF bytes.

        **A `200` does not mean the document opened.** The variants that report a problem — `error`, `dwg_binary`, `office_error`, `revit_binary`, `3d_error`, `unsupported` — are also returned with `200`. This is a defect kept for compatibility: every client branches on `type`, so correcting the status is a coordinated change on both sides.

        Requires the `document:read` permission.
      operationId: getViewerPayload
      parameters:
      - description: Identifier of the document to open.
        example: 1180
        in: path
        name: documentId
        required: true
        schema:
          type:
          - integer
          format: int64
          example: null
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewerPayload"
                example: null
            application/pdf:
              schema:
                format: binary
                example: null
            image/*:
              schema:
                format: binary
                example: null
            model/gltf-binary:
              schema:
                format: binary
                example: null
          description: "What the viewer needs, or a described reason it cannot open\
            \ the document."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: No document with that id is visible to the caller.
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Get what the viewer needs to open a document
      tags:
      - Viewer
...
  /api/viewer/{documentId}/pdf:
    get:
      description: |-
        The bytes of the document's current version, served inline for the viewer to render.

        Not cached: processing replaces the bytes behind this URL, so a cached response would show a version that no longer exists. The viewer passes the version in the query string for the same reason.

        Requires the `document:read` permission.
      operationId: getPdfBytes
      parameters:
      - description: Identifier of the document.
        example: 1180
        in: path
        name: documentId
        required: true
        schema:
          type:
          - integer
          format: int64
          example: null
      - description: Version the client believes it is fetching. Present so a browser
          does not serve a superseded copy from cache; the current version is always
          what is returned.
        example: 3
        in: query
        name: v
        required: false
        schema:
          type:
          - integer
          format: int32
          example: null
      responses:
        "200":
          content:
            application/pdf:
              schema:
                format: binary
                example: null
          description: The PDF's bytes.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No such document, or its file is missing from storage."
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The document is not a PDF.
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Stream a PDF's bytes
      tags:
      - Viewer
...
components:
  schemas:
...
    ConverterStatusResponse:
      description: "Whether the document conversion service is reachable. The viewer\
        \ asks before offering conversion-dependent actions, so it can say why something\
        \ is unavailable rather than failing when it is tried."
      example: null
      properties:
        running:
          type:
          - boolean
          description: Whether the conversion service answered.
          example: true
      required:
      - running
...
    ProblemDetail:
      type:
      - object
      description: "RFC 9457 Problem Details. The single error shape returned by every\
        \ endpoint, for every status."
      example: null
      properties:
        detail:
          type:
          - string
          description: "What happened and what to do next, written for a person to\
            \ read. Never contains a stack trace, SQL, or an internal type name."
          example: 1 field is invalid.
        instance:
          type:
          - string
          format: uri-reference
          description: Path of the request that failed.
          example: /api/projects
        invalidFields:
          type:
          - array
          description: "Present on `422` validation failures. Names every field that\
            \ was rejected, so a client can attach each message to the control it\
            \ belongs to rather than to the form as a whole."
          example: null
          items:
            type:
            - object
            description: One field the request was rejected for.
            example: null
            properties:
              field:
                type:
                - string
                description: "Path of the rejected field, as it appears in the request\
                  \ body."
                example: name
              message:
                type:
                - string
                description: "What is wrong with the value, phrased for the person\
                  \ who typed it."
                example: must not be blank
            required:
            - field
            - message
        status:
          type:
          - integer
          format: int32
          description: "HTTP status code, repeated here so a logged body is self-contained."
          example: 422
          maximum: 599
          minimum: 400
        timestamp:
          type:
          - string
          format: date-time
          description: "When the failure was recorded, UTC."
          example: 2026-02-17T09:41:12.884Z
        title:
          type:
          - string
          description: Short summary of the problem type. Does not vary between occurrences.
          example: Validation failed
        traceId:
          type:
          - string
          description: "Identifies this request in the server logs, and is also returned\
            \ in the `X-Trace-Id` header. Quote it when reporting a problem."
          example: 4f8a1c2e9b7d6a5f3e2d1c0b9a8f7e6d
          pattern: "^[0-9a-f]{32}$"
        type:
          type:
          - string
          format: uri-reference
          description: "Stable identifier for this kind of problem, relative to the\
            \ API host. Safe to branch on — unlike `detail`, its value does not change\
            \ with wording."
          example: /problems/validation-failed
      required:
      - type
      - title
      - status
      - detail
      - traceId
...
  securitySchemes:
    apiKey:
      description: "Machine credential, sent as `X-API-Key`. Shown once at creation\
        \ and stored only as a SHA-256 digest, so a lost key is replaced rather than\
        \ recovered. Scoped, expiring, and rate-limited per key."
      in: header
      name: X-API-Key
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      description: "Signed JSON Web Token from `POST /api/auth/login`, sent as `Authorization:\
        \ Bearer <token>`. The token carries the subject and the tenant; the tenant\
        \ claim is authoritative and is not overridable by any request parameter."
      scheme: bearer
      type: http
    oauth2:
      description: Authorization code with PKCE for user-facing clients; client credentials
        for machine clients. The implicit flow is not offered.
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          refreshUrl: /oauth2/token
          scopes:
            annotation:read: Read markup on a document
            annotation:write: "Create, update and resolve markup"
            document:process: Run operations that rewrite a document
            document:read: Read document metadata and content
            document:write: "Upload, replace and delete documents"
            project:read: List and read projects
            project:write: "Create, update and delete projects"
            signature:read: Read and verify digital signatures
            signature:write: Apply a digital signature to a document version
          tokenUrl: /oauth2/token
        clientCredentials:
          scopes:
            annotation:read: Read markup on a document
            annotation:write: "Create, update and resolve markup"
            document:process: Run operations that rewrite a document
            document:read: Read document metadata and content
            document:write: "Upload, replace and delete documents"
            project:read: List and read projects
...